rename everything to mantaray
