+++ mantaray/colors.py
