+++ mantaray/images/sources.txt
